+++ core.impl/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/impl/jmx/DistributedJmxProxyFactoryImpl.java
@@ -16,6 +16,9 @@ package com.ericsson.bss.cassandra.ecchronos.core.impl.jmx;
 
 import com.datastax.oss.driver.api.core.metadata.Node;
 import com.ericsson.bss.cassandra.ecchronos.connection.DistributedJmxConnectionProvider;
+import com.ericsson.bss.cassandra.ecchronos.core.impl.jmx.http.ClientRegisterResponse;
+import com.ericsson.bss.cassandra.ecchronos.core.impl.jmx.http.NotificationListenerResponse;
+import com.ericsson.bss.cassandra.ecchronos.core.impl.jmx.http.NotificationRegisterResponse;
 import com.ericsson.bss.cassandra.ecchronos.core.jmx.DistributedJmxProxy;
 import com.ericsson.bss.cassandra.ecchronos.core.jmx.DistributedJmxProxyFactory;
 import com.ericsson.bss.cassandra.ecchronos.core.table.TableReference;
@@ -23,11 +26,26 @@ import com.ericsson.bss.cassandra.ecchronos.core.table.TableReference;
 import com.ericsson.bss.cassandra.ecchronos.data.sync.EccNodesSync;
 import com.ericsson.bss.cassandra.ecchronos.utils.enums.sync.NodeStatus;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import java.net.URI;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
 import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.UUID;
 import java.util.ArrayList;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.TimeUnit;
+import javax.management.Notification;
 import javax.management.ObjectName;
 import java.io.IOException;
 
@@ -40,16 +58,23 @@ import javax.management.NotificationListener;
 import javax.management.ReflectionException;
 import javax.management.openmbean.CompositeData;
 import javax.management.remote.JMXConnector;
+import com.google.common.util.concurrent.ThreadFactoryBuilder;
 
+import org.jolokia.client.J4pClient;
+import org.jolokia.client.request.J4pExecRequest;
+
+import org.jolokia.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * A factory creating JMX proxies to Cassandra.
  */
-@SuppressWarnings({"PMD.ClassWithOnlyPrivateConstructorsShouldBeFinal", "checkstyle:finalclass"})
+@SuppressWarnings({"PMD.ClassWithOnlyPrivateConstructorsShouldBeFinal", "checkstyle:finalclass", "PMD.GodClass"})
 public class DistributedJmxProxyFactoryImpl implements DistributedJmxProxyFactory
 {
+    private static final long DEFAULT_RUN_DELAY_IN_MS = 100;
+    private static final String CLIENT_ID_PROPERTY = "clientID";
     private static final Logger LOG = LoggerFactory.getLogger(DistributedJmxProxyFactoryImpl.class);
     private static final String SS_OBJ_NAME = "org.apache.cassandra.db:type=StorageService";
     private static final String RS_OBJ_NAME = "org.apache.cassandra.db:type=RepairService";
@@ -62,12 +87,14 @@ public class DistributedJmxProxyFactoryImpl implements DistributedJmxProxyFactor
     private final DistributedJmxConnectionProvider myDistributedJmxConnectionProvider;
     private final Map<UUID, Node> nodesMap;
     private final EccNodesSync eccNodesSync;
+    private final boolean isJolokiaEnabled;
 
     private DistributedJmxProxyFactoryImpl(final Builder builder)
     {
         myDistributedJmxConnectionProvider = builder.myDistributedJmxConnectionProvider;
         nodesMap = builder.myNodesMap;
         eccNodesSync = builder.myEccNodesSync;
+        isJolokiaEnabled = builder.isJolokiaEnabled;
     }
 
     @Override
@@ -78,7 +105,8 @@ public class DistributedJmxProxyFactoryImpl implements DistributedJmxProxyFactor
             return new InternalDistributedJmxProxy(
                     myDistributedJmxConnectionProvider,
                     nodesMap,
-                    eccNodesSync);
+                    eccNodesSync,
+                    isJolokiaEnabled);
         }
         catch (MalformedObjectNameException e)
         {
@@ -90,14 +118,28 @@ public class DistributedJmxProxyFactoryImpl implements DistributedJmxProxyFactor
     {
         private final DistributedJmxConnectionProvider myDistributedJmxConnectionProvider;
         private final Map<UUID, Node> myNodesMap;
+        private final Map<NotificationListener, String> myJolokiaRelationshipListeners = new HashMap<>();
+
+        private final ScheduledExecutorService myNotificationExecutor = Executors.newSingleThreadScheduledExecutor(
+                new ThreadFactoryBuilder().setNameFormat("NotificationRefresher-%d").build());
+
+        private final Map<UUID, Map<String, String>> myClientIdMap = new ConcurrentHashMap<>();
+        private final Map<UUID, Map<String, NotificationListener>> myNodeListenersMap = new ConcurrentHashMap<>();
+        private final Map<UUID, Map<String, ScheduledFuture<?>>> myNotificationMonitors = new ConcurrentHashMap<>();
+
+        private final boolean isJolokiaEnabled;
         private final EccNodesSync myEccNodesSync;
         private final ObjectName myStorageServiceObject;
         private final ObjectName myRepairServiceObject;
 
+        private final ObjectMapper objectMapper = new ObjectMapper();
+        private final HttpClient client = HttpClient.newHttpClient();
+
         private InternalDistributedJmxProxy(
                 final DistributedJmxConnectionProvider distributedJmxConnectionProvider,
                 final Map<UUID, Node> nodesMap,
-                final EccNodesSync eccNodesSync
+                final EccNodesSync eccNodesSync,
+                final boolean jolokiaEnabled
         ) throws MalformedObjectNameException
         {
             myDistributedJmxConnectionProvider = distributedJmxConnectionProvider;
@@ -105,6 +147,218 @@ public class DistributedJmxProxyFactoryImpl implements DistributedJmxProxyFactor
             myEccNodesSync = eccNodesSync;
             myStorageServiceObject = new ObjectName(SS_OBJ_NAME);
             myRepairServiceObject = new ObjectName(RS_OBJ_NAME);
+            isJolokiaEnabled = jolokiaEnabled;
+        }
+
+        private void startNotificationMonitor(final UUID nodeID, final String notificationID)
+        {
+            ScheduledFuture<?> future = myNotificationExecutor.scheduleWithFixedDelay(
+                    new NotificationRunTask(nodeID, notificationID), 0, DEFAULT_RUN_DELAY_IN_MS, TimeUnit.MILLISECONDS);
+
+            synchronized (myNotificationMonitors)
+            {
+                myNotificationMonitors
+                        .computeIfAbsent(nodeID, k -> new ConcurrentHashMap<>())
+                        .put(notificationID, future);
+            }
+        }
+
+        private final class NotificationRunTask implements Runnable
+        {
+            private final UUID myNodeID;
+            private final String myNotificationID;
+            private final Set<Integer> notificationController = new HashSet<>();
+
+            private NotificationRunTask(final UUID nodeID, final String notificationID)
+            {
+                myNodeID = nodeID;
+                myNotificationID = notificationID;
+            }
+
+            @Override
+            public void run()
+            {
+                try
+                {
+                    JSONObject response = checkForNotifications(myNodeID, myNotificationID);
+                    NotificationListenerResponse notificationListenerResponse = objectMapper.readValue(response.toJSONString(),
+                            NotificationListenerResponse.class);
+
+                    if (notificationListenerResponse.getValue() != null)
+                    {
+                        List<NotificationListenerResponse.Notification> notifications =
+                                notificationListenerResponse.getValue().getNotifications();
+
+                        for (NotificationListenerResponse.Notification notificationObj : notifications)
+                        {
+                            Notification notification = new Notification(
+                                    notificationObj.getType(),
+                                    notificationObj.getSource(),
+                                    notificationObj.getTimeStamp(),
+                                    notificationObj.getMessage()
+                            );
+                            notification.setUserData(notificationObj.getUserData());
+                            if (!notificationController.contains(notification.hashCode()))
+                            {
+                                notificationController.add(notification.hashCode());
+                                synchronized (myNodeListenersMap)
+                                {
+                                    NotificationListener listener = myNodeListenersMap.get(myNodeID).get(myNotificationID);
+                                    if (listener != null)
+                                    {
+                                        listener.handleNotification(notification, null);
+                                    }
+                                    else
+                                    {
+                                        LOG.warn("Listener not found for node {} and notificationID {}", myNodeID, myNotificationID);
+                                    }
+                                }
+                            }
+                        }
+                    }
+
+                }
+                catch (Exception e)
+                {
+                    LOG.error("Error monitoring notifications for node {} and notificationID {}: {}", myNodeID, myNotificationID, e.getMessage());
+                }
+            }
+        }
+
+        private void registerClientId(final UUID nodeID)
+        {
+            try
+            {
+                String url = mountJolokiaBaseURL(nodeID) + "/notification/register";
+
+                HttpRequest request = HttpRequest.newBuilder().uri(URI.create(url)).GET().build();
+
+                HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
+
+                ClientRegisterResponse clientRegisterResponse = objectMapper.readValue(response.body(),
+                        ClientRegisterResponse.class);
+
+                Map<String, String> properties = new HashMap<>();
+                properties.put(CLIENT_ID_PROPERTY, clientRegisterResponse.getValue().getId());
+                properties.put("store", clientRegisterResponse.getValue().getBackend().getPull().getStore());
+
+                myClientIdMap.put(nodeID, properties);
+            }
+            catch (IOException | InterruptedException e)
+            {
+                LOG.error("Unable to register Jolokia Client in node with ID {} because of {}", nodeID, e.getMessage());
+            }
+        }
+
+        private String registerJolokiaNotification(final UUID nodeID) throws IOException, InterruptedException
+        {
+            String url = mountJolokiaBaseURL(nodeID) + "/notification";
+
+            HttpRequest request = HttpRequest.newBuilder()
+                    .uri(URI.create(url)).POST(HttpRequest.BodyPublishers.ofString(jolokiaCreateNotificationOptions(nodeID))).build();
+            HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
+
+            NotificationRegisterResponse notificationRegisterResponse = objectMapper.readValue(response.body(),
+                    NotificationRegisterResponse.class);
+
+            return notificationRegisterResponse.getValue();
+        }
+
+        private void removeJolokiaNotification(final UUID nodeID, final String notificationID)
+        {
+            String url = mountJolokiaBaseURL(nodeID) + "/notification";
+            HttpRequest request = HttpRequest.newBuilder()
+                    .uri(URI.create(url)).POST(HttpRequest.BodyPublishers.ofString(
+                        jolokiaRemoveNotificationOptions(nodeID, notificationID))).build();
+            try
+            {
+                client.send(request, HttpResponse.BodyHandlers.ofString());
+            }
+            catch (IOException | InterruptedException e)
+            {
+                LOG.error("Error trying to remove NotificationListener with ID {} in node {}, because of {}",
+                        notificationID, nodeID, e.getMessage());
+            }
+        }
+
+        private String jolokiaCreateNotificationOptions(final UUID nodeID)
+        {
+            Map<String, Object> params = new HashMap<>();
+            params.put("type", "notification");
+            params.put("command", "add");
+            params.put("client", myClientIdMap.get(nodeID).get(CLIENT_ID_PROPERTY));
+            params.put("mode", "pull");
+            params.put("mbean", SS_OBJ_NAME);
+            List<String> filter = List.of(
+                    "progress",
+                    "jmx.remote.connection.lost.notifications",
+                    "jmx.remote.connection.failed",
+                    "jmx.remote.connection.closed"
+            );
+            params.put("filter", filter);
+
+            try
+            {
+                return objectMapper.writeValueAsString(params);
+            }
+            catch (JsonProcessingException e)
+            {
+                LOG.error("Unable to serialize notification options for node {} because of {}", nodeID, e.getMessage());
+            }
+            return "";
+        }
+
+        private String jolokiaRemoveNotificationOptions(final UUID nodeID, final String notificationID)
+        {
+            Map<String, Object> params = new HashMap<>();
+            params.put("type", "notification");
+            params.put("command", "remove");
+            params.put("client", myClientIdMap.get(nodeID).get(CLIENT_ID_PROPERTY));
+            params.put("handle", notificationID);
+
+            try
+            {
+                return objectMapper.writeValueAsString(params);
+            }
+            catch (JsonProcessingException e)
+            {
+                LOG.error("Unable to serialize Jolokia Notification Options for node {} because of {}", nodeID,
+                        e.getMessage());
+            }
+            return "";
+        }
+
+        private J4pClient mountJolokiaClient(final UUID nodeID)
+        {
+            return new J4pClient(mountJolokiaBaseURL(nodeID));
+        }
+
+        private String mountJolokiaBaseURL(final UUID nodeID)
+        {
+            String host = myNodesMap.get(nodeID).getBroadcastRpcAddress().get().getHostString();
+            return "http://" + host + ":8778/jolokia";
+        }
+
+        private JSONObject checkForNotifications(final UUID nodeID, final String notificationID)
+        {
+            String operation = "pull";
+            try
+            {
+                synchronized (myClientIdMap)
+                {
+                    J4pExecRequest execRequest = new J4pExecRequest(
+                            myClientIdMap.get(nodeID).get("store"),
+                            operation, myClientIdMap.get(nodeID).get(CLIENT_ID_PROPERTY),
+                            notificationID);
+
+                    return mountJolokiaClient(nodeID).execute(execRequest).asJSONObject();
+                }
+            }
+            catch (Exception e)
+            {
+                LOG.error("Error checking notifications for node {} and notificationID {}: {}", nodeID, notificationID, e.getMessage());
+            }
+            return new JSONObject();
         }
 
         @Override
@@ -123,9 +377,27 @@ public class DistributedJmxProxyFactoryImpl implements DistributedJmxProxyFactor
                 try
                 {
                     nodeConnection.addConnectionNotificationListener(listener, null, null);
-                    nodeConnection.getMBeanServerConnection().addNotificationListener(myStorageServiceObject, listener, null, null);
+                    if (isJolokiaEnabled)
+                    {
+                        registerClientId(nodeID);
+
+                        String jolokiaNotificationID = registerJolokiaNotification(nodeID);
+
+                        synchronized (myNodeListenersMap)
+                        {
+                            myNodeListenersMap.computeIfAbsent(nodeID, k -> new ConcurrentHashMap<>()).put(jolokiaNotificationID, listener);
+                        }
+
+                        myJolokiaRelationshipListeners.put(listener, jolokiaNotificationID);
+
+                        startNotificationMonitor(nodeID, jolokiaNotificationID);
+                    }
+                    else
+                    {
+                        nodeConnection.getMBeanServerConnection().addNotificationListener(myStorageServiceObject, listener, null, null);
+                    }
                 }
-                catch (InstanceNotFoundException | IOException e)
+                catch (InstanceNotFoundException | IOException | InterruptedException e)
                 {
                     LOG.error("Unable to add StorageService listener in node {} with because of {}", nodeID, e.getMessage());
                 }
@@ -295,7 +567,23 @@ public class DistributedJmxProxyFactoryImpl implements DistributedJmxProxyFactor
                 try
                 {
                     nodeConnection.removeConnectionNotificationListener(listener);
-                    nodeConnection.getMBeanServerConnection().removeNotificationListener(myStorageServiceObject, listener);
+                    if (isJolokiaEnabled)
+                    {
+                        String jolokiaNotificationID = myJolokiaRelationshipListeners.get(listener);
+                        removeJolokiaNotification(nodeID, jolokiaNotificationID);
+                        synchronized (myNotificationMonitors)
+                        {
+                            myNotificationMonitors.get(nodeID).get(jolokiaNotificationID).cancel(true);
+                            myNotificationMonitors.get(nodeID).remove(jolokiaNotificationID);
+                        }
+                        myNodeListenersMap.get(nodeID).remove(jolokiaNotificationID);
+                        myJolokiaRelationshipListeners.remove(listener);
+                    }
+                    else
+                    {
+                        nodeConnection.getMBeanServerConnection().removeNotificationListener(myStorageServiceObject, listener);
+                    }
+
                 }
                 catch (InstanceNotFoundException | ListenerNotFoundException | IOException e)
                 {
@@ -482,6 +770,7 @@ public class DistributedJmxProxyFactoryImpl implements DistributedJmxProxyFactor
             LOG.error("Unable to get connection with node {}, marking as UNREACHABLE", nodeID);
             myEccNodesSync.updateNodeStatus(NodeStatus.UNAVAILABLE, myNodesMap.get(nodeID).getDatacenter(), nodeID);
         }
+
     }
 
     public static Builder builder()
@@ -494,6 +783,7 @@ public class DistributedJmxProxyFactoryImpl implements DistributedJmxProxyFactor
         private DistributedJmxConnectionProvider myDistributedJmxConnectionProvider;
         private Map<UUID, Node> myNodesMap;
         private EccNodesSync myEccNodesSync;
+        private boolean isJolokiaEnabled = false;
 
         /**
          * Build with JMX connection provider.
@@ -528,6 +818,18 @@ public class DistributedJmxProxyFactoryImpl implements DistributedJmxProxyFactor
         public Builder withEccNodesSync(final EccNodesSync eccNodesSync)
         {
             myEccNodesSync = eccNodesSync;
+            return this;
+        }
+
+        /**
+         * Build with Jolokia Agent.
+         *
+         * @param jolokiaEnabled Define if Jolokia Client must be used.
+         * @return Builder
+         */
+        public Builder withJolokiaEnabled(final boolean jolokiaEnabled)
+        {
+            isJolokiaEnabled = jolokiaEnabled;
             return this;
         }